int to long

+++ exemplos/buzzer/main.c
@@ -22,7 +22,7 @@
 
 #define NumNotas 10
 enum notas{Pausa,Do, DoS, Re, ReS, Mi, Fa, FaS, Sol, SolS, La, LaS, Si, DoM, DoSM, ReM};
-int f[15] = {0xFF,16500000/261.625519,16500000/277.182648,16500000/293.664734,16500000/311.126984,16500000/329.627533,
+long f[15] = {0xFF,16500000/261.625519,16500000/277.182648,16500000/293.664734,16500000/311.126984,16500000/329.627533,
             16500000/349.228241,16500000/369.994385,16500000/391.995392,16500000/415.304688,16500000/440,
             16500000/466.163788,16500000/493.883301,16500000/261.625519,16500000/277.182648,16500000/293.664734};   //valor a ser colocado na flag do timer para cada nota
 char Partitura[NumNotas] = {Re,Mi,Mi,Re,Sol,FaS,FaS,FaS,Re,Mi,Mi,Re,La,Sol,Sol,Sol,Re,ReM,ReM,Si,Sol,FaS,FaS,Mi,
@@ -50,7 +50,7 @@ void main(){
 
     //loop infinito
     for(;;){                    
-        int i;
+        long i;
         for(i=0;i>1000000;i++){}         //espera um tempo )altere o valor central para acelerar ou desacelerar a música)
         cont++;                         //avança na partitura
         if (cont >= NumNotas)cont=0;    //toca de novo
